+++ tools/__init__.py
@@ -2,16 +2,24 @@
 
 from .image_tools import GenerateImageTool
 from .video_tools import GenerateVideoTool
+from .video_task_tools import QueryVideoTaskTool, DownloadVideoTool
 from .music_tools import GenerateMusicTool
+from .music_cover_tools import MusicCoverTool
 from .search_tools import WebSearchTool
 from .vision_tools import DescribeImageTool
 from .quota_tools import CheckQuotaTool
+from .speech_tools import SpeechSynthesizeTool, ListVoicesTool
 
 __all__ = [
     "GenerateImageTool",
     "GenerateVideoTool",
+    "QueryVideoTaskTool",
+    "DownloadVideoTool",
     "GenerateMusicTool",
+    "MusicCoverTool",
     "WebSearchTool",
     "DescribeImageTool",
     "CheckQuotaTool",
+    "SpeechSynthesizeTool",
+    "ListVoicesTool",
 ]